Applying token authentication
