+++ start/kernel.ts
@@ -9,7 +9,7 @@
 |
 */
 
-import Server from '@ioc:Adonis/Core/Server'
+import Server from "@ioc:Adonis/Core/Server";
 
 /*
 |--------------------------------------------------------------------------
@@ -20,9 +20,7 @@ import Server from '@ioc:Adonis/Core/Server'
 | are defined for every HTTP requests.
 |
 */
-Server.middleware.register([
-  () => import('@ioc:Adonis/Core/BodyParser'),
-])
+Server.middleware.register([() => import("@ioc:Adonis/Core/BodyParser")]);
 
 /*
 |--------------------------------------------------------------------------
@@ -41,4 +39,5 @@ Server.middleware.register([
 |
 */
 Server.middleware.registerNamed({
-})
+  auth: () => import("App/Middleware/Auth"),
+});
