+++ start/routes.ts
@@ -1,50 +1,26 @@
 import Route from "@ioc:Adonis/Core/Route";
 
-//verifica se a API está online
+//Verifica se a API está online
 Route.get("/", async () => {
-  return "A api-estoque está online!";
+  return { apiStorage: "online" };
 });
 
-//gera o token de acesso do usuário
-Route.post("/login", async ({ auth, request, response }) => {
-  const email = request.input("email");
-  const password = request.input("password");
+//Gera o token de acesso do usuário
+Route.post("/login", "SessionsController.store");
 
-  try {
-    const token = await auth.use("api").attempt(email, password);
-    return token;
-  } catch {
-    return response.unauthorized("Invalid credentials");
-  }
-});
-
-//verifica se o usuário tem ou não token de acesso
-Route.get("dashboard", async ({ auth }) => {
-  await auth.use("api").authenticate();
-  return `Olá ${auth.user?.name}, você está autenticado(a)!`;
-});
-
-//revoga o token
-Route.post("/logout", async ({ auth }) => {
-  try {
-    await auth.use("api").authenticate();
-    await auth.use("api").revoke();
-    return {
-      revoked: true,
-    };
-  } catch (error) {
-    console.log(error);
-  }
-});
+//Verifica se o usuário tem ou não token de acesso
+Route.get("/session", "SessionsController.show").middleware("auth:api");
 
 //Users routes
+Route.post("/", "UsersController.store").prefix("users");
 Route.group(() => {
   Route.get("/", "UsersController.index");
   Route.get("/:id", "UsersController.show");
-  Route.post("/", "UsersController.store");
   Route.put("/:id", "UsersController.update");
   Route.delete("/:id", "UsersController.destroy");
-}).prefix("users");
+})
+  .middleware("auth:api")
+  .prefix("users");
 
 //Categories routes
 Route.group(() => {
@@ -53,7 +29,9 @@ Route.group(() => {
   Route.post("/", "CategoriesController.store");
   Route.put("/:id", "CategoriesController.update");
   Route.delete("/:id", "CategoriesController.destroy");
-}).prefix("categories");
+})
+  .middleware("auth:api")
+  .prefix("categories");
 
 //Suppliers routes
 Route.group(() => {
@@ -62,4 +40,6 @@ Route.group(() => {
   Route.post("/", "SuppliersController.store");
   Route.put("/:id", "SuppliersController.update");
   Route.delete("/:id", "SuppliersController.destroy");
-}).prefix("suppliers");
+})
+  .middleware("auth:api")
+  .prefix("suppliers");
